create folder to solutiors,move py files to folder,modified README
