+++ README.md
@@ -2,7 +2,7 @@
 
 |                       Описание                        |               Реализация                |
 |:-----------------------------------------------------:|:---------------------------------------:|
-| [Переход на новую вкладку](tasks/switch_to_window.md) | [switch_to_window](switch_to_window.py) |
+| [Переход на новую вкладку](tasks/switch_to_window.md) | [switch_to_window](solutions/switch_to_window.py) |
 
 ---
 fill_the_huge_form.py https://stepik.org/lesson/138920/step/7?unit=196194
