+++ solutions/check_registration_form.py
