+++ solutions/get_attribute.py
