+++ solutions/switch_to_window.py
