+++ solutions/working_with_dropdown_list.py
